[PDS-89] fix:관리자 자원 예약 목록 조회 컨트롤러 이전
