+++ src/main/java/com/example/pladialmserver/booking/service/BookingService.java
@@ -11,6 +11,7 @@ import com.example.pladialmserver.booking.repository.resourceBooking.ResourceBoo
 import com.example.pladialmserver.global.entity.BookingStatus;
 import com.example.pladialmserver.global.exception.BaseException;
 import com.example.pladialmserver.global.exception.BaseResponseCode;
+import com.example.pladialmserver.resource.dto.response.AdminResourceRes;
 import com.example.pladialmserver.user.entity.Role;
 import com.example.pladialmserver.user.entity.User;
 import lombok.RequiredArgsConstructor;
@@ -256,6 +257,21 @@ public class BookingService {
     public void returnBookingResourceByAdmin(User user, Long resourceBookingId) {
         ResourceBooking resourceBooking = checkResourceBookingAuthentication(user, resourceBookingId, Role.ADMIN);
         returnBookingResource(resourceBooking);
+    }
+
+    /**
+     * 관리자 자원 예약 목록을 조회
+     */
+    public Page<AdminResourceRes> getBookingResources(Pageable pageable) {
+        Pageable sortedByDateAsc = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(),
+                Sort.by(Sort.Order.asc("startDate")));
+
+        Page<ResourceBooking> resourceBookings=resourceBookingRepository.findByStatusIn(
+                Arrays.asList(BookingStatus.BOOKED, BookingStatus.USING,BookingStatus.WAITING),
+                sortedByDateAsc
+        );
+
+        return resourceBookings.map(AdminResourceRes::toDto);
     }
 
 }
